split files `models.py` and `test.py`
In order ease development for all developers and management of the library
split each class into its dedicated file.

split the file `test.py` to simplify management of tests.
One file per class being tested.

! This is a breaking change ! user will be required to update their code
in order to match library new layout.
