+++ tox.ini
@@ -4,4 +4,4 @@ envlist = py36,py37,py38,py39
 [testenv]
 passenv = GS_RECORD_MODE GS_CREDS_FILENAME
 deps = -rtest-requirements.txt
-commands = nosetests -vv tests/test.py
+commands = nosetests -vv tests/
